// __generated__/__types__.tsx
export type Maybe<T> = T;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
};

export type Analytics = {
  __typename?: 'Analytics';
  google_analytics?: Maybe<Scalars['String']>;
  umami_id?: Maybe<Scalars['String']>;
  simple_analytics?: Maybe<Scalars['String']>;
  plausible_data_domain?: Maybe<Scalars['String']>;
};

export type Author = {
  __typename?: 'Author';
  id: Scalars['Int'];
  email: Scalars['String'];
  username: Scalars['String'];
  name: Scalars['String'];
  social?: Maybe<Social>;
  bio?: Maybe<Scalars['String']>;
  occupation?: Maybe<Scalars['String']>;
  company_name?: Maybe<Scalars['String']>;
  role?: Maybe<Role>;
  permissions?: Maybe<Array<Permissions>>;
  avatar?: Maybe<Scalars['String']>;
  verified?: Maybe<Scalars['Boolean']>;
  accessToken?: Maybe<Scalars['String']>;
  analytics_id?: Maybe<Scalars['Int']>;
  analytics_uuid?: Maybe<Scalars['String']>;
};
// ...
export type Image = {
  __typename?: 'Image';
  src?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
  height?: Maybe<Scalars['Int']>;
};
// ...
export type LetterpadError = {
  message: Scalars['String'];
};
// ...
export enum NavigationType {
  Tag = 'tag',
  Page = 'page',
  Custom = 'custom'
}

export enum Permissions {
  ManageOwnPosts = 'MANAGE_OWN_POSTS',
  ReadOnlyPosts = 'READ_ONLY_POSTS',
  ManageAllPosts = 'MANAGE_ALL_POSTS',
  ManageUsers = 'MANAGE_USERS',
  ManageSettings = 'MANAGE_SETTINGS'
}

export type Post = {
  __typename?: 'Post';
  id: Scalars['Int'];
  title: Scalars['String'];
  html?: Maybe<Scalars['String']>;
  html_draft?: Maybe<Scalars['String']>;
  author?: Maybe<Author>;
  excerpt?: Maybe<Scalars['String']>;
  cover_image: Image;
  type: PostTypes;
  status: PostStatusOptions;
  featured: Scalars['Boolean'];
  slug?: Maybe<Scalars['String']>;
  createdAt: Scalars['Date'];
  publishedAt?: Maybe<Scalars['Date']>;
  scheduledAt?: Maybe<Scalars['Date']>;
  updatedAt: Scalars['Date'];
  reading_time?: Maybe<Scalars['String']>;
  tags?: Maybe<Array<Tags>>;
};

export type PostCountsByStatus = {
  __typename?: 'PostCountsByStatus';
  published: Scalars['Int'];
  drafts: Scalars['Int'];
};

export type PostError = LetterpadError & {
  __typename?: 'PostError';
  message: Scalars['String'];
};

export type PostFilters = {
  id?: Maybe<Scalars['Int']>;
  slug?: Maybe<Scalars['String']>;
  featured?: Maybe<Scalars['Boolean']>;
  previewHash?: Maybe<Scalars['String']>;
  status?: Maybe<PostStatusOptions>;
  type?: Maybe<PostTypes>;
};

export type PostResponse = Post | PostError;

export enum PostStatusOptions {
  Published = 'published',
  Draft = 'draft',
  Trashed = 'trashed'
}

export enum PostTypes {
  Post = 'post',
  Page = 'page'
}

export type PostsFilters = {
  id?: Maybe<Scalars['Int']>;
  type?: Maybe<PostTypes>;
  slug?: Maybe<Scalars['String']>;
  featured?: Maybe<Scalars['Boolean']>;
  previewHash?: Maybe<Scalars['String']>;
  status?: Maybe<PostStatusOptions>;
  author?: Maybe<Scalars['String']>;
  tagSlug?: Maybe<Scalars['String']>;
  tag?: Maybe<Scalars['String']>;
  cursor?: Maybe<Scalars['Int']>;
  page?: Maybe<Scalars['Int']>;
  offset?: Maybe<Scalars['Int']>;
  limit?: Maybe<Scalars['Int']>;
  sortBy?: Maybe<SortBy>;
};

export type PostsNode = {
  __typename?: 'PostsNode';
  count: Scalars['Int'];
  rows: Array<Post>;
};

export type PostsResponse = PostsNode | PostError;
// ...
export type QueryPostsArgs = {
  filters?: Maybe<PostsFilters>;
};
// ...
export enum SortBy {
  Asc = 'asc',
  Desc = 'desc'
}
// ...
export type Tags = {
  __typename?: 'Tags';
  name: Scalars['String'];
  slug: Scalars['String'];
  posts?: Maybe<PostsResponse>;
};
// ...
export type ResolverTypeWrapper<T> = Promise<T> | T;
// ...
export type ResolversTypes = {
// ...
  Post: ResolverTypeWrapper<Post>;
  PostCountsByStatus: ResolverTypeWrapper<PostCountsByStatus>;
  PostError: ResolverTypeWrapper<PostError>;
  PostFilters: PostFilters;
  PostResponse: ResolversTypes['Post'] | ResolversTypes['PostError'];
  PostStatusOptions: PostStatusOptions;
  PostTypes: PostTypes;
  PostsFilters: PostsFilters;
  PostsNode: ResolverTypeWrapper<PostsNode>;
  PostsResponse: ResolversTypes['PostsNode'] | ResolversTypes['PostError'];
// ...
};